hima: tfa9895: Fix build for 64bit

+++ amplifier/tfa9895.c
@@ -274,7 +274,7 @@ static int tfa9895_read_reg(struct tfa9895_amp_t *amp, uint8_t reg,
     }
 
     reg_val[0] = 2;
-    reg_val[1] = (unsigned int) &buf;
+    reg_val[1] = (unsigned long) &buf;
     /* unsure why the first byte is skipped */
     buf[0] = 0;
     buf[1] = reg;
@@ -284,7 +284,7 @@ static int tfa9895_read_reg(struct tfa9895_amp_t *amp, uint8_t reg,
     }
 
     reg_val[0] = 2;
-    reg_val[1] = (unsigned int) &buf;
+    reg_val[1] = (unsigned long) &buf;
     if ((ret = ioctl(amp->fd, TPA9895_READ_CONFIG, &reg_val)) != 0) {
         ALOGE("ioctl %d failed, ret = %d", TPA9895_READ_CONFIG, ret);
         goto read_reg_err;
@@ -309,7 +309,7 @@ static int tfa9895_write_reg(struct tfa9895_amp_t *amp, uint8_t reg,
     }
 
     reg_val[0] = 4;
-    reg_val[1] = (unsigned int) &buf;
+    reg_val[1] = (unsigned long) &buf;
     /* unsure why the first byte is skipped */
     buf[0] = 0;
     buf[1] = reg;
@@ -338,7 +338,7 @@ static int tfa9895_read(struct tfa9895_amp_t *amp, int addr, uint8_t *buf,
     }
 
     reg_val[0] = 2;
-    reg_val[1] = (unsigned int) &reg_buf;
+    reg_val[1] = (unsigned long) &reg_buf;
     /* unsure why the first byte is skipped */
     reg_buf[0] = 0;
     reg_buf[1] = (0xFF & addr);
@@ -348,7 +348,7 @@ static int tfa9895_read(struct tfa9895_amp_t *amp, int addr, uint8_t *buf,
     }
 
     reg_val[0] = len;
-    reg_val[1] = (unsigned int) &kernel_buf;
+    reg_val[1] = (unsigned long) &kernel_buf;
     if ((ret = ioctl(amp->fd, TPA9895_READ_CONFIG, &reg_val)) != 0) {
         ALOGE("ioctl %d failed, ret = %d", TPA9895_READ_CONFIG, ret);
         goto read_err;
@@ -372,7 +372,7 @@ static int tfa9895_write(struct tfa9895_amp_t *amp, int addr,
     }
 
     reg_val[0] = len + 2;
-    reg_val[1] = (unsigned int) &ioctl_buf;
+    reg_val[1] = (unsigned long) &ioctl_buf;
     /* unsure why the first byte is skipped */
     ioctl_buf[0] = 0;
     ioctl_buf[1] = (0xFF & addr);
